Create initial market state. #13

+++ src/main/java/ch/uzh/ifi/seal/soprafs18/game/cards/MovingCard.java
@@ -22,7 +22,7 @@ import java.util.List;
 public class MovingCard extends Card{
 
     @JsonCreator
-    public MovingCard(String name, int coinValue, int coinCost, int strength, int depth, COLOR[] colors){
+    public MovingCard(String name, float coinValue, int coinCost, int strength, int depth, COLOR[] colors){
         super(name, coinValue, coinCost);
         this.strength = strength;
         this.depth = depth;
